DEL-260: В dto изменил тип price с Int на BigDecimal

+++ feature-route/impl/src/client/kotlin/orderchanges/presentation/compose/model/OrderChangedValuesUiModel.kt
@@ -2,11 +2,12 @@ package orderchanges.presentation.compose.model
 
 import cargotype.domain.model.CargoType
 import extras.presentation.model.ExtrasUiModel
+import java.math.BigDecimal
 
 data class OrderChangedValuesUiModel(
     val boxes: Int,
     val pallets: Int,
     val cargoType: CargoType,
-    val price: Int,
+    val price: BigDecimal,
     val extras: List<ExtrasUiModel>,
 )
